+++ NEWS.md
@@ -1,5 +1,7 @@
 # eemR 0.1.2
 
+- `eem_extract()` has now an argument `ignore_case` (#10) to specify if the regular expression search should ignore sample name case (TRUE) or not (FALSE).
+
 # eemR 0.1.1
 
 - Fixing minimal R version to run the package (R >= 3.2.1)